PR #1069 review rework
